# api/openapi/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiError": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"details": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"error",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DirectoryValidation": {
				"description": "Result of validating a directory for use as data storage.",
				"properties": {
					"isValid": {
						"type": "boolean",
						"description": "Whether the directory can be used"
					},
					"exists": {
						"type": "boolean",
						"description": "Whether the directory exists"
					},
					"isWritable": {
						"type": "boolean",
						"description": "Whether the directory is writable"
					},
					"hasExistingData": {
						"type": "boolean",
						"description": "Whether data already exists at this location"
					},
					"existingFiles": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "List of existing data files found"
					},
					"error": {
						"type": "string",
						"description": "Error message if not valid"
					}
				},
				"required": [
					"isValid",
					"exists",
					"isWritable",
					"hasExistingData",
					"existingFiles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateDirectoryRequest": {
				"properties": {
					"path": {
						"type": "string"
					}
				},
				"required": [
					"path"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MigrationResult": {
				"description": "Result of a data migration operation.",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Whether the migration succeeded"
					},
					"entityFilesCopied": {
						"type": "number",
						"format": "double",
						"description": "Number of entity files copied"
					},
					"monthFilesCopied": {
						"type": "number",
						"format": "double",
						"description": "Number of month files copied"
					},
					"filesCopied": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "List of files that were copied"
					},
					"error": {
						"type": "string",
						"description": "Error message if failed"
					},
					"sourceDir": {
						"type": "string",
						"description": "Source directory"
					},
					"destDir": {
						"type": "string",
						"description": "Destination directory"
					}
				},
				"required": [
					"success",
					"entityFilesCopied",
					"monthFilesCopied",
					"filesCopied",
					"sourceDir",
					"destDir"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MigrateDataRequest": {
				"properties": {
					"sourceDir": {
						"type": "string"
					},
					"destDir": {
						"type": "string"
					},
					"mode": {
						"type": "string",
						"enum": [
							"copy",
							"fresh",
							"use_existing"
						]
					}
				},
				"required": [
					"sourceDir",
					"destDir",
					"mode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SwitchDirectoryRequest": {
				"properties": {
					"newDirectory": {
						"type": "string"
					}
				},
				"required": [
					"newDirectory"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "BudgetForFun API",
		"version": "1.0.0",
		"description": "BudgetForFun API - Personal budget tracking application",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "BudgetForFun Team"
		}
	},
	"paths": {
...
		"/api/settings/validate-directory": {
			"post": {
				"operationId": "ValidateDirectory",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DirectoryValidation"
								}
							}
						}
					}
				},
				"description": "Validate a directory for use as data storage",
				"summary": "Validate directory",
				"tags": [
					"Settings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateDirectoryRequest"
							}
						}
					}
				}
			}
		},
		"/api/settings/migrate-data": {
			"post": {
				"operationId": "MigrateData",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MigrationResult"
								}
							}
						}
					},
					"400": {
						"description": "Migration failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Migrate data to a new directory",
				"summary": "Migrate data",
				"tags": [
					"Settings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MigrateDataRequest"
							}
						}
					}
				}
			}
		},
		"/api/settings/switch-directory": {
			"post": {
				"operationId": "SwitchDirectory",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Switch failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Switch to a different data directory",
				"summary": "Switch directory",
				"tags": [
					"Settings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SwitchDirectoryRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}